refactor: errorCheck 클래스 분리
