+++ docs/README.md
@@ -14,5 +14,5 @@
 4. 사용자가 잘못된 값 입력 시 에러 처리
     - 입력이 number 타입이 아닐 때
     - 플레이어의 수가 2명 이하일 때
-5. 예외사항 처리
 
+5. 예외사항 처리
